Fixed allreduce, but...
Fixed allreduce by setting correct dimensions, but the numbers being returned don't seem to be right. Ran twice with 100 and 10000 points, 100 points definitely not good as the variance and error are of the same magnitude as the integral, 10000 points fixes this but the values seem to be smaller than they should be?